backend/scout_agent/requirements.txt
fastapi
uvicorn
google-adk
google-cloud-firestore
google-cloud-core
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
google-cloud-secret-manager
google-cloud-spanner
google-cloud-speech
google-cloud-texttospeech
google-cloud-trace
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-appengine-logging
googlemaps
google-auth
google-auth-httplib2
google-api-python-client
vertexai
google-genai
pandas
python-dotenv
lxml
numpy
pydantic
gradio
firebase-admin
httpx
requests
twilio
python-multipart
audioop-lts
a2a-sdk